fix: 총 수익률 출력 화면 수정

+++ src/main/java/lotto/domain/User.java
@@ -22,7 +22,7 @@ public class User {
         long totalProfit = statistics.getTotalProfit();
 
         float profitPercent = (float)totalProfit / (float) purchaseMoney * 100;
-        profitPercent = Math.round(profitPercent * 10) / 10;
+        profitPercent = (float) Math.round(profitPercent * 10) / 10;
 
         return profitPercent;
     }
